oip5: repair proto package paths and dependencies
