+++ modules/oip5/record_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/golang/protobuf/proto"
 
 	"github.com/oipwg/oip/datastore"
+	"github.com/oipwg/oip/modules/oip"
 )
 
 // import (
