+++ proto/build_proto.sh
@@ -7,7 +7,11 @@ protoc --go_out=$GOPATH/src multipart.proto pubkey.proto signedMessage.proto txi
 echo "Building oip5 proto files"
 cd $GOPATH/src/github.com/oipwg/oip/proto/oip5
 protoc --go_out=$GOPATH/src -I=. -I=$GOPATH/src/github.com/oipwg/oip/proto/oip \
-       edit.proto NormalizeRecord.proto NormalizeRecord.proto oip5.proto Record.proto RecordTemplateProto.proto
+  edit.proto NormalizeRecord.proto NormalizeRecord.proto oip5.proto Record.proto RecordTemplateProto.proto
+
+echo "Building oip5 template proto files"
+cd $GOPATH/src/github.com/oipwg/oip/proto/oip5/templates
+protoc --go_out=$GOPATH/src tmpl_433C2783.proto
 
 echo "Building historian proto files"
 cd $GOPATH/src/github.com/oipwg/oip/proto/historian
