+++ proto/oip/multipart.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package oip;
+package oipProto;
 
 option go_package = "github.com/oipwg/oip/modules/oip;oip";
